requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
selenium==4.15.2
Pillow==10.0.1
Werkzeug==2.3.7
requests==2.31.0

